support managing volumes with SDK
