+++ sky/clouds/cudo.py
@@ -12,7 +12,7 @@ from sky.utils import resources_utils
 if typing.TYPE_CHECKING:
     # Renaming to avoid shadowing variables.
     from sky import resources as resources_lib
-    from sky.volumes import volume as volume_lib
+    from sky.utils import volume as volume_lib
 
 _CREDENTIAL_FILES = [
     # credential files for Cudo,
