+++ sky/clouds/do.py
@@ -14,7 +14,7 @@ from sky.utils import resources_utils
 
 if typing.TYPE_CHECKING:
     from sky import resources as resources_lib
-    from sky.volumes import volume as volume_lib
+    from sky.utils import volume as volume_lib
 
 _CREDENTIAL_FILE = 'config.yaml'
