+++ sky/clouds/gcp.py
@@ -29,7 +29,7 @@ from sky.utils import ux_utils
 if typing.TYPE_CHECKING:
     from sky import resources
     from sky.utils import status_lib
-    from sky.volumes import volume as volume_lib
+    from sky.utils import volume as volume_lib
 
 logger = sky_logging.init_logger(__name__)
