+++ sky/clouds/hyperbolic.py
@@ -13,7 +13,7 @@ from sky.utils.resources_utils import DiskTier
 
 if typing.TYPE_CHECKING:
     from sky import resources as resources_lib
-    from sky.volumes import volume as volume_lib
+    from sky.utils import volume as volume_lib
 
 
 @registry.CLOUD_REGISTRY.register
