+++ sky/clouds/paperspace.py
@@ -14,7 +14,7 @@ if typing.TYPE_CHECKING:
     import requests
 
     from sky import resources as resources_lib
-    from sky.volumes import volume as volume_lib
+    from sky.utils import volume as volume_lib
 else:
     requests = adaptors_common.LazyImport('requests')
