+++ sky/volumes/client/sdk.py
@@ -26,6 +26,12 @@ logger = sky_logging.init_logger(__name__)
 @annotations.client_api
 def apply(volume: volume_lib.Volume) -> server_common.RequestId:
     """Creates or registers a volume.
+
+    Args:
+        volume: The volume to apply.
+
+    Returns:
+        The request ID of the apply request.
     """
     body = payloads.VolumeApplyBody(name=volume.name,
                                     volume_type=volume.type,
@@ -45,7 +51,11 @@ def apply(volume: volume_lib.Volume) -> server_common.RequestId:
 @server_common.check_server_healthy_or_start
 @annotations.client_api
 def ls() -> server_common.RequestId:
-    """Lists all volumes."""
+    """Lists all volumes.
+
+    Returns:
+        The request ID of the list request.
+    """
     response = requests.get(f'{server_common.get_server_url()}/volumes',
                             cookies=server_common.get_api_cookie_jar())
     return server_common.get_request_id(response)
@@ -56,7 +66,14 @@ def ls() -> server_common.RequestId:
 @server_common.check_server_healthy_or_start
 @annotations.client_api
 def delete(names: List[str]) -> server_common.RequestId:
-    """Deletes a volume."""
+    """Deletes volumes.
+
+    Args:
+        names: List of volume names to delete.
+
+    Returns:
+        The request ID of the delete request.
+    """
     body = payloads.VolumeDeleteBody(names=names)
     response = requests.post(f'{server_common.get_server_url()}/volumes/delete',
                              json=json.loads(body.model_dump_json()),
